requirements.txt
# Minimal requirements for TTS + RVC voice conversion
edge-tts
numpy
soundfile
librosa
torch
faiss-cpu
soxr
noisereduce
pedalboard
wget
transformers
torchaudio
resampy
torchcrepe
# Optional: torchfcpe if using fcpe F0 extraction

# FastAPI dependencies
fastapi
uvicorn[standard]
pydantic

# Testing dependencies
requests

# New dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-chroma>=0.0.5
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0